Revert "feat(chat): pinned messages (#201)"
This reverts commit 34a232e49309955267d2f56ba61a17e9e13360fc.
