--- src/lib/commands/built-in/hide-pin.ts
@@ -1,12 +1,0 @@
-import { defineCommand } from "../util";
-
-export default defineCommand({
-	provider: "Built-in",
-	name: "hide-pin",
-	description: "Hide the pinned message in chat",
-	async exec(_, channel) {
-		if (channel.chat.pinned) {
-			channel.chat.pinned.hidden = true;
-		}
-	},
-});
