--- src/lib/commands/built-in/unpin.ts
@@ -1,12 +1,0 @@
-import { defineCommand } from "../util";
-
-export default defineCommand({
-	provider: "Built-in",
-	name: "unpin",
-	description: "Unpin the currently pinned message in the channel",
-	modOnly: true,
-	async exec(_, channel) {
-		await channel.chat.fetchPinned();
-		await channel.chat.pinned?.unpin();
-	},
-});
