+++ src/lib/handlers/eventsub/channel-chat-user-message-hold.ts
@@ -10,9 +10,10 @@ export default defineHandler({
 
 		data.message.message_id = data.message_id;
 
-		const message = UserMessage.from(channel, {
-			message: data.message,
-			sender: data,
+		const message = UserMessage.from(channel, data.message, {
+			id: data.user_id,
+			login: data.user_login,
+			name: data.user_name,
 		});
 
 		message.autoMod = {
