+++ src/lib/handlers/eventsub/channel-suspicious-user-message.ts
@@ -8,9 +8,10 @@ export default defineHandler({
 		const channel = app.channels.get(data.broadcaster_user_id);
 		if (!channel) return;
 
-		const message = UserMessage.from(channel, {
-			message: data.message,
-			sender: data,
+		const message = UserMessage.from(channel, data.message, {
+			id: data.user_id,
+			login: data.user_login,
+			name: data.user_name,
 		});
 
 		if (message.viewer) {
