+++ src/lib/menus/message-menu.ts
@@ -29,18 +29,6 @@ export async function createMessageMenu(message: UserMessage) {
 	items.push(copy, separator, reply);
 
 	if (message.channel.isMod) {
-		const pin = await MenuItem.new({
-			id: "pin",
-			text: "Pin",
-			enabled: !message.deleted && message.actionable && !message.pinned,
-			async action() {
-				await message.pin();
-				await message.channel.chat.fetchPinned();
-			},
-		});
-
-		items.push(pin);
-
 		const deleteMsg = await MenuItem.new({
 			id: "delete",
 			text: "Delete",
