+++ src/lib/twitch/api.ts
@@ -3,63 +3,6 @@ export interface DropReason {
 	message: string;
 }
 
-export interface TextFragment {
-	type: "text";
-	text: string;
-}
-
-export interface CheermoteFragment {
-	type: "cheermote";
-	text: string;
-	cheermote: {
-		prefix: string;
-		bits: number;
-		tier: number;
-	};
-}
-
-export interface EmoteFragment {
-	type: "emote";
-	text: string;
-	emote: {
-		id: string;
-		emote_set_id: string;
-		owner_id?: string;
-		formats?: string[];
-	};
-}
-
-export interface MentionFragment {
-	type: "mention";
-	text: string;
-	user_id: string;
-	user_login: string;
-	user_name: string;
-}
-
-export type Fragment = TextFragment | CheermoteFragment | EmoteFragment | MentionFragment;
-
-export interface StructuredMessage {
-	message_id: string;
-	text: string;
-	fragments: Fragment[];
-}
-
-export interface PinnedMessage {
-	message_id: string;
-	broadcaster_id: string;
-	sender_user_id: string;
-	sender_user_login: string;
-	sender_user_name: string;
-	pinned_by_user_id: string;
-	pinned_by_user_login: string;
-	pinned_by_user_name: string;
-	message: StructuredMessage;
-	starts_at: string;
-	expires_at: string | null;
-	updated_at: string;
-}
-
 export interface SentMessage {
 	message_id: string;
 	is_sent: boolean;
